fix: rename test-setup.ts to exclude from Convex deployment
Fixes production Convex deployment by excluding test setup file.

- Renamed `convex/test-setup.ts` to `convex/test-setup.test.ts`
- Updated all imports in test files (chats, files, messages, promptTemplates, users, lib/rateLimiter)
- Convex automatically excludes `*.test.ts` files from deployment

Root cause: `test-setup.ts` was being deployed to production, but it contains imports
from `@convex-dev/rate-limiter/dist/component/*` which aren't exported by the package.
These imports only work locally but fail in production bundling.

By renaming to `test-setup.test.ts`, Convex will exclude it from deployment automatically,
preventing the module resolution errors.

Tested locally:
- ✅ Build passes with no errors
- ✅ Test files correctly import from renamed file
