+++ apps/server/convex/messages.test.ts
@@ -16,7 +16,7 @@ import { convexTest } from 'convex-test';
 import schema from './schema';
 import { api } from './_generated/api';
 import type { Id } from './_generated/dataModel';
-import { modules, rateLimiter } from './test-setup';
+import { modules, rateLimiter } from './test-setup.test';
 
 // Helper to create convex test instance with components registered
 function createConvexTest() {
